requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.1
Werkzeug==3.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.7



